Rename useAuthUser file
