+++ src/withUser.tsx
@@ -90,7 +90,7 @@ const withAuthUser: WithAuthUser =
       var React = require('react')
       var { useEffect, useCallback, useMemo } = require('react')
       var { useRouter } = require('next/router')
-      var { AuthUserContext } = require('src/useAuthUser')
+      var { AuthUserContext } = require('src/useUser')
       /* eslint-enable global-require, no-var, vars-on-top, @typescript-eslint/no-var-requires */
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (e: any) {
